+++ app/policies/project_item_policy.rb
@@ -22,9 +22,12 @@ class ProjectItemPolicy < ApplicationPolicy
       (a.value_schema[:type] == "array" || a.type == "select" ? { a.id => [] } : a.id)
       # TODO handle other attribute value types
     }
-    [:service_id, :project_id, :affiliation_id, :customer_typology,
-     :access_reason, :additional_information, :user_group_name,
-     :project_name, :project_website_url, :company_name,
-     :company_website_url, property_values: attributes]
+
+    [
+      :service_id, :project_id, :affiliation_id, :customer_typology,
+      :access_reason, :additional_information, :user_group_name,
+      :project_name, :project_website_url, :company_name, :research_area_id,
+      :company_website_url, property_values: attributes
+    ]
   end
 end