+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ## [Unreleased]
 
 ### Added
+- User needs to select "Research area" while ordering service (@mkasztelnik)
 
 ### Changed